feat:合并搜索栏与导航条，实现搜索功能的搜索关键字保存，但不能实现依靠参数的状态跳转。
