--- src/components/Searchbar.js
@@ -1,72 +1,0 @@
-import * as React from "react";
-import { styled, alpha } from "@mui/material/styles";
-import { Route, Routes, useNavigate } from "react-router-dom";
-import InputBase from "@mui/material/InputBase";
-import SearchIcon from "@mui/icons-material/Search";
-import Searchpage from "../views/Searchpage";
-
-const inputKeyUp = (e) => {
-  if (e.keyCode === 13) {
-    // alert(e.target.value);
-    console.log(e.target.value);
-    <Routes>
-      <Route path="/search" element={<Searchpage />} />
-    </Routes>;
-  }
-};
-
-const Search = styled("div")(({ theme }) => ({
-  position: "relative",
-  borderRadius: theme.shape.borderRadius,
-  backgroundColor: alpha(theme.palette.common.white, 0.15),
-  "&:hover": {
-    backgroundColor: alpha(theme.palette.common.white, 0.25),
-  },
-  marginRight: theme.spacing(2),
-  marginLeft: 0,
-  width: "100%",
-  [theme.breakpoints.up("sm")]: {
-    marginLeft: theme.spacing(3),
-    width: "auto",
-  },
-}));
-
-const SearchIconWrapper = styled("div")(({ theme }) => ({
-  padding: theme.spacing(0, 2),
-  height: "100%",
-  position: "absolute",
-  pointerEvents: "none",
-  display: "flex",
-  alignItems: "center",
-  justifyContent: "center",
-}));
-
-const StyledInputBase = styled(InputBase)(({ theme }) => ({
-  color: "inherit",
-  "& .MuiInputBase-input": {
-    padding: theme.spacing(1, 1, 1, 0),
-    // vertical padding + font size from searchIcon
-    paddingLeft: `calc(1em + ${theme.spacing(4)})`,
-    transition: theme.transitions.create("width"),
-    width: "100%",
-    [theme.breakpoints.up("md")]: {
-      width: "20ch",
-    },
-  },
-}));
-
-export default function Searchbar({ search, onSearch }) {
-  return (
-    <Search onKeyUp={inputKeyUp}>
-      <SearchIconWrapper>
-        <SearchIcon />
-      </SearchIconWrapper>
-      <StyledInputBase
-        value={search}
-        onChange={onSearch}
-        placeholder="发现更多…"
-        inputProps={{ "aria-label": "search" }}
-      />
-    </Search>
-  );
-}
